[services, common] Remove useless files and fix some warnings

+++ src/emu/common/src/bytepair.cpp
@@ -163,7 +163,7 @@ namespace eka2l1 {
             return static_cast<int>(dest - static_cast<uint8_t *>(destination));
             
         done_dest:
-            return dest - static_cast<uint8_t *>(destination);
+            return static_cast<int>(dest - static_cast<uint8_t *>(destination));
 
             return 1;
         }
